Implement authorization using Policy

+++ app/Http/Controllers/QuestionsController.php
@@ -7,6 +7,17 @@ use Illuminate\Http\Request;
 
 class QuestionsController extends Controller
 {
+
+    /**
+     * Constructeur pour appeler un middleware pour interdire l'accès aux non loggués à tout sauf aux méthodes index () et show ()
+     *
+     * @return void
+    */
+    public function __construct ()
+    {
+        $this->middleware('auth', ['except' => ['index', 'show']]);
+    }
+
     /**
      * Display a listing of the resource.
      *
@@ -60,6 +71,7 @@ class QuestionsController extends Controller
      */
     public function edit (Question $question)
     {
+        $this->authorize("update", $question);
         return view('questions.edit', compact('question'));
     }
     /**
@@ -71,8 +83,10 @@ class QuestionsController extends Controller
      */
     public function update (AskQuestionRequest $request, Question $question)
     {
+        $this->authorize("update", $question);
+    
         $question->update($request->only('title', 'body'));
-
+    
         return redirect()->route('questions.index')->with('success', "votre question a bien été mise à jour !");
     }
     /**
@@ -83,8 +97,9 @@ class QuestionsController extends Controller
      */
     public function destroy (Question $question)
     {
+        $this->authorize("delete", $question);
+    
         $question->delete();
-
-        return redirect('/questions')->with('success', "votre question dont le titre était : « $question->title » a bien été supprimée !");
+        return redirect()->route('questions.index')->with('success', "votre question a bien été supprimée !");
     }
 }
